Update lnmp from OneinStack

+++ include/GraphicsMagick.sh
@@ -9,39 +9,59 @@
 #       https://github.com/oneinstack/oneinstack
 
 Install_GraphicsMagick() {
-  pushd ${oneinstack_dir}/src > /dev/null
-  tar xzf GraphicsMagick-${graphicsmagick_ver}.tar.gz
-  pushd GraphicsMagick-${graphicsmagick_ver}
-  ./configure --prefix=${gmagick_install_dir} --enable-shared --enable-static
-  make -j ${THREAD} && make install
-  popd
-  rm -rf GraphicsMagick-${graphicsmagick_ver}
-  popd
+  if [ -d "${gmagick_install_dir}" ]; then
+    echo "${CWARNING}GraphicsMagick already installed! ${CEND}"
+  else
+    pushd ${oneinstack_dir}/src > /dev/null
+    tar xzf GraphicsMagick-${graphicsmagick_ver}.tar.gz
+    pushd GraphicsMagick-${graphicsmagick_ver} > /dev/null
+    ./configure --prefix=${gmagick_install_dir} --enable-shared --enable-static
+    make -j ${THREAD} && make install
+    popd > /dev/null
+    rm -rf GraphicsMagick-${graphicsmagick_ver}
+    popd > /dev/null
+  fi
+}
+
+Uninstall_GraphicsMagick() {
+  if [ -d "${gmagick_install_dir}" ]; then
+    rm -rf ${gmagick_install_dir}
+    echo; echo "${CMSG}GraphicsMagick uninstall completed${CEND}"
+  fi
 }
 
 Install_pecl-gmagick() {
-  pushd ${oneinstack_dir}/src > /dev/null
   if [ -e "${php_install_dir}/bin/phpize" ]; then
+    pushd ${oneinstack_dir}/src > /dev/null
     phpExtensionDir=`${php_install_dir}/bin/php-config --extension-dir`
     if [ "`${php_install_dir}/bin/php -r 'echo PHP_VERSION;' | awk -F. '{print $1}'`" == '7' ]; then
-      tar xzf gmagick-${gmagick_for_php7_ver}.tgz
-      pushd gmagick-${gmagick_for_php7_ver}
-    else
       tar xzf gmagick-${gmagick_ver}.tgz
-      pushd gmagick-${gmagick_ver}
+      pushd gmagick-${gmagick_ver} > /dev/null
+    else
+      tar xzf gmagick-${gmagick_oldver}.tgz
+      pushd gmagick-${gmagick_oldver} > /dev/null
     fi
     export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig
     ${php_install_dir}/bin/phpize
     ./configure --with-php-config=${php_install_dir}/bin/php-config --with-gmagick=${gmagick_install_dir}
     make -j ${THREAD} && make install
-    popd
+    popd > /dev/null
     if [ -f "${phpExtensionDir}/gmagick.so" ]; then
       echo 'extension=gmagick.so' > ${php_install_dir}/etc/php.d/03-gmagick.ini
       echo "${CSUCCESS}PHP gmagick module installed successfully! ${CEND}"
-      rm -rf gmagick-${gmagick_for_php7_ver} gmagick-${gmagick_ver}
+      rm -rf gmagick-${gmagick_ver} gmagick-${gmagick_oldver}
     else
       echo "${CFAILURE}PHP gmagick module install failed, Please contact the author! ${CEND}"
     fi
+    popd > /dev/null
   fi
-  popd
+}
+
+Uninstall_pecl-gmagick() {
+  if [ -e "${php_install_dir}/etc/php.d/03-gmagick.ini" ]; then
+    rm -f ${php_install_dir}/etc/php.d/03-gmagick.ini
+    echo; echo "${CMSG}PHP gmagick module uninstall completed${CEND}"
+  else
+    echo; echo "${CWARNING}PHP gmagick module does not exist! ${CEND}"
+  fi
 }
