+++ include/ImageMagick.sh
@@ -9,27 +9,38 @@
 #       https://github.com/oneinstack/oneinstack
 
 Install_ImageMagick() {
-  pushd ${oneinstack_dir}/src > /dev/null
-  tar xzf ImageMagick-${imagemagick_ver}.tar.gz
-  pushd ImageMagick-${imagemagick_ver}
-  ./configure --prefix=${imagick_install_dir} --enable-shared --enable-static
-  make -j ${THREAD} && make install
-  popd
-  rm -rf ImageMagick-${imagemagick_ver}
-  popd
+  if [ -d "${imagick_install_dir}" ]; then
+    echo "${CWARNING}ImageMagick already installed! ${CEND}"
+  else
+    pushd ${oneinstack_dir}/src > /dev/null
+    tar xzf ImageMagick-${imagemagick_ver}.tar.gz
+    pushd ImageMagick-${imagemagick_ver} > /dev/null
+    ./configure --prefix=${imagick_install_dir} --enable-shared --enable-static
+    make -j ${THREAD} && make install
+    popd > /dev/null
+    rm -rf ImageMagick-${imagemagick_ver}
+    popd > /dev/null
+  fi
+}
+
+Uninstall_ImageMagick() {
+  if [ -d "${imagick_install_dir}" ]; then
+    rm -rf ${imagick_install_dir} 
+    echo; echo "${CMSG}ImageMagick uninstall completed${CEND}"
+  fi
 }
 
 Install_pecl-imagick() {
-  pushd ${oneinstack_dir}/src > /dev/null
   if [ -e "${php_install_dir}/bin/phpize" ]; then
+    pushd ${oneinstack_dir}/src > /dev/null
     phpExtensionDir=`${php_install_dir}/bin/php-config --extension-dir`
     tar xzf imagick-${imagick_ver}.tgz
-    pushd imagick-${imagick_ver}
+    pushd imagick-${imagick_ver} > /dev/null
     export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig
     ${php_install_dir}/bin/phpize
     ./configure --with-php-config=${php_install_dir}/bin/php-config --with-imagick=${imagick_install_dir}
     make -j ${THREAD} && make install
-    popd
+    popd > /dev/null
     if [ -f "${phpExtensionDir}/imagick.so" ]; then
       echo 'extension=imagick.so' > ${php_install_dir}/etc/php.d/03-imagick.ini
       echo "${CSUCCESS}PHP imagick module installed successfully! ${CEND}"
@@ -37,6 +48,15 @@ Install_pecl-imagick() {
     else
       echo "${CFAILURE}PHP imagick module install failed, Please contact the author! ${CEND}"
     fi
+    popd > /dev/null
   fi
-  popd
+}
+
+Uninstall_pecl-imagick() {
+  if [ -e "${php_install_dir}/etc/php.d/03-imagick.ini" ]; then
+    rm -f ${php_install_dir}/etc/php.d/03-imagick.ini
+    echo; echo "${CMSG}PHP imagick module uninstall completed${CEND}"
+  else
+    echo; echo "${CWARNING}PHP imagick module does not exist! ${CEND}"
+  fi
 }
