+++ include/zendopcache.sh
@@ -9,27 +9,29 @@
 #       https://github.com/oneinstack/oneinstack
 
 Install_ZendOPcache() {
-  pushd ${oneinstack_dir}/src > /dev/null
-  phpExtensionDir=$(${php_install_dir}/bin/php-config --extension-dir)
-  PHP_detail_ver=$(${php_install_dir}/bin/php -r 'echo PHP_VERSION;')
-  PHP_main_ver=${PHP_detail_ver%.*}
-  if [[ "${PHP_main_ver}" =~ ^5.[3-4]$ ]]; then
-    tar xzf zendopcache-${zendopcache_ver}.tgz
-    pushd zendopcache-${zendopcache_ver}
-  else
-    tar xzf php-${PHP_detail_ver}.tar.gz
-    pushd php-${PHP_detail_ver}/ext/opcache
-  fi
-
-  ${php_install_dir}/bin/phpize
-  ./configure --with-php-config=${php_install_dir}/bin/php-config
-  make -j ${THREAD} && make install
-  popd
-  if [ -f "${phpExtensionDir}/opcache.so" ]; then
-    # write opcache configs
+  if [ -e "${php_install_dir}/bin/phpize" ]; then
+    pushd ${oneinstack_dir}/src > /dev/null
+    phpExtensionDir=$(${php_install_dir}/bin/php-config --extension-dir)
+    PHP_detail_ver=$(${php_install_dir}/bin/php -r 'echo PHP_VERSION;')
+    PHP_main_ver=${PHP_detail_ver%.*}
     if [[ "${PHP_main_ver}" =~ ^5.[3-4]$ ]]; then
-      # For php 5.3 5.4
-      cat > ${php_install_dir}/etc/php.d/02-opcache.ini << EOF
+      tar xzf zendopcache-${zendopcache_ver}.tgz
+      pushd zendopcache-${zendopcache_ver} > /dev/null
+    else
+      src_url=http://www.php.net/distributions/php-${PHP_detail_ver}.tar.gz && Download_src
+      tar xzf php-${PHP_detail_ver}.tar.gz
+      pushd php-${PHP_detail_ver}/ext/opcache > /dev/null
+    fi
+
+    ${php_install_dir}/bin/phpize
+    ./configure --with-php-config=${php_install_dir}/bin/php-config
+    make -j ${THREAD} && make install
+    popd > /dev/null
+    if [ -f "${phpExtensionDir}/opcache.so" ]; then
+      # write opcache configs
+      if [[ "${PHP_main_ver}" =~ ^5.[3-4]$ ]]; then
+        # For php 5.3 5.4
+        cat > ${php_install_dir}/etc/php.d/02-opcache.ini << EOF
 [opcache]
 zend_extension=${phpExtensionDir}/opcache.so
 opcache.enable=1
@@ -42,9 +44,10 @@ opcache.fast_shutdown=1
 opcache.enable_cli=1
 ;opcache.optimization_level=0
 EOF
-    else
-      # For php 5.5+
-      cat > ${php_install_dir}/etc/php.d/02-opcache.ini << EOF
+        rm -rf zendopcache-${zendopcache_ver}
+      else
+        # For php 5.5+
+        cat > ${php_install_dir}/etc/php.d/02-opcache.ini << EOF
 [opcache]
 zend_extension=opcache.so
 opcache.enable=1
@@ -61,12 +64,22 @@ opcache.fast_shutdown=1
 opcache.consistency_checks=0
 ;opcache.optimization_level=0
 EOF
-    fi
+      fi
 
-    echo "${CSUCCESS}PHP OPcache module installed successfully! ${CEND}"
-    rm -rf zendopcache-${zendopcache_ver} php-${PHP_detail_ver}
-  else
-    echo "${CFAILURE}PHP OPcache module install failed, Please contact the author! ${CEND}"
+      echo "${CSUCCESS}PHP opcache module installed successfully! ${CEND}"
+      rm -rf php-${PHP_detail_ver}
+    else
+      echo "${CFAILURE}PHP opcache module install failed, Please contact the author! ${CEND}"
+    fi
+    popd > /dev/null
   fi
-  popd
+}
+
+Uninstall_ZendOPcache() {
+  if [ -e "${php_install_dir}/etc/php.d/02-opcache.ini" ]; then
+    rm -f ${php_install_dir}/etc/php.d/02-opcache.ini
+    echo; echo "${CMSG}PHP opcache module uninstall completed${CEND}"
+  else
+    echo; echo "${CWARNING}PHP opcache module does not exist! ${CEND}"
+  fi
 }
